グラフ操作のロジックをカスタムhooksに切り出し
